--- src/app/reporting/loading.tsx
@@ -1,5 +1,0 @@
-import { LoadingSkeleton } from "@/features/loading-skeleton"
-
-export default function Loading(): JSX.Element {
-  return <LoadingSkeleton />
-}
