--- src/app/reporting/page.tsx
@@ -1,5 +1,0 @@
-import { Reporting, ReportingProp } from "@/features/reporting/reporting"
-
-export default function Home(props: ReportingProp): JSX.Element {
-  return <Reporting {...props} />
-}
